requirements.txt
pydantic>=2.0.0
typer>=0.12.0
PyYAML>=6.0
ollama>=0.3.0
pypdf>=4.0.0
python-docx>=1.1.0
pytesseract>=0.3.10
Pillow>=10.0.0
python-pptx>=0.6.0
